requirements.txt
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.11
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.2
